--- src/app/services/api-objects/_thumb-max-json.ts
@@ -1,10 +1,0 @@
-import { JsonObject, JsonProperty } from 'json2typescript';
-
-@JsonObject
-export class ThumbMaxJson {
-    @JsonProperty('nx', Number)
-    public nx: number = undefined;
-
-    @JsonProperty('ny', Number)
-    public ny: number = undefined;
-}
